requirements.txt
selenium~=3.141.0
pytest~=6.2.4
allure-pytest==2.9.42
